+++ acceptance-tests/mssql-db/mssql_db_subsume_test.go
@@ -1,10 +1,15 @@
-package mssql_test
+package mssql_db_test
 
 import (
 	"acceptancetests/apps"
 	"acceptancetests/helpers"
 	"acceptancetests/helpers/cf"
 	"acceptancetests/helpers/random"
+	"os/exec"
+	"strings"
+	"time"
+
+	"github.com/onsi/gomega/gexec"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
@@ -39,11 +44,18 @@ var _ = Describe("MSSQL DB Subsume", func() {
 		By("fetching the Azure resource ID of the database")
 		resource := fetchResourceID("db", masbDBName, metadata.PreProvisionedSQLServer)
 
-		By("reconfiguring the CSB with DB server details")
-		serverTag := reconfigureCSBWithMASBServerDetails()
+		By("Create CSB with DB server details")
+		serverTag := random.Name(random.WithMaxLength(10))
+		creds := getMASBServerDetails(serverTag)
+
+		serviceBroker := helpers.CreateBroker(
+			helpers.BrokerWithPrefix("csb-mssql-db"),
+			helpers.BrokerWithEnv(helpers.EnvVar{Name: "MSSQL_DB_SERVER_CREDS", Value: creds}),
+		)
+		defer serviceBroker.Delete()
 
 		By("subsuming the database")
-		csbServiceInstance := helpers.CreateServiceFromBroker("csb-azure-mssql-db", "subsume", helpers.DefaultBroker().Name, subsumeDBParams(resource, serverTag))
+		csbServiceInstance := helpers.CreateServiceFromBroker("csb-azure-mssql-db", "subsume", serviceBroker.Name, subsumeDBParams(resource, serverTag))
 		defer csbServiceInstance.Delete()
 
 		By("purging the MASB service instance")
@@ -79,9 +91,8 @@ func subsumeDBParams(resource, serverTag string) interface{} {
 	}
 }
 
-func reconfigureCSBWithMASBServerDetails() string {
-	tag := random.Name(random.WithMaxLength(10))
-	creds := map[string]interface{}{
+func getMASBServerDetails(tag string) map[string]interface{} {
+	return map[string]interface{}{
 		tag: map[string]string{
 			"server_name":           metadata.PreProvisionedSQLServer,
 			"server_resource_group": metadata.ResourceGroup,
@@ -89,12 +100,6 @@ func reconfigureCSBWithMASBServerDetails() string {
 			"admin_password":        metadata.PreProvisionedSQLPassword,
 		},
 	}
-
-	helpers.SetBrokerEnvAndRestart(
-		helpers.EnvVar{Name: "MSSQL_DB_SERVER_CREDS", Value: creds},
-	)
-
-	return tag
 }
 
 func masbServerConfig(dbName string) interface{} {
@@ -103,4 +108,12 @@ func masbServerConfig(dbName string) interface{} {
 		"sqldbName":     dbName,
 		"resourceGroup": metadata.ResourceGroup,
 	}
+}
+
+func fetchResourceID(kind, name, server string) string {
+	command := exec.Command("az", "sql", kind, "show", "--name", name, "--server", server, "--resource-group", metadata.ResourceGroup, "--query", "id", "-o", "tsv")
+	session, err := gexec.Start(command, GinkgoWriter, GinkgoWriter)
+	Expect(err).NotTo(HaveOccurred())
+	Eventually(session, time.Minute).Should(gexec.Exit(0))
+	return strings.TrimSpace(string(session.Out.Contents()))
 }
